Changes for P.R. Beamforming #30
Separated out coefficients generator class into separate class file.
Removed @jit added for coeff generator as causing numba deprecation warnings (code actually ran slower with it)
Fixed real-imag swap (and wrong dot product) in unit test (CPU version).

TODO:
Add buffer slot for coefficients rather than coeffs being member of class in instattiation.
Fix Sgemm memory issue. It was noted that sometimes pytest fails some tests when Sgemm_test is inluded (causes kernel version which follows to fail as well).
