+++ beamformer/test/beamform_Sgemm_test.py
@@ -22,118 +22,8 @@ import numpy as np
 import pytest
 import test_parameters
 from beamforming import matrix_multiply
+from coeff_generator import CoeffGenerator
 from katsdpsigproc import accel
-from numba import jit
-
-
-class CoeffGenerator:
-    """Class for generating coefficients for testing purposes.
-
-    Note: The coefficients for this feature will be supplied by CAM and are not
-    generated by CBF. This code can removed when suitable coefficients are available.
-
-    Parameters
-    ----------
-    n_batches: int
-        The number of matrices to be reordered, a single data matrix = one batch.
-    n_channels: int
-        The number of frequency channels to be processed.
-    n_blocks: int
-        The number of blocks that each channels set of samples are divided into.
-    samples_per_block: int
-        The number of time samples to be processed per block.
-    n_ants: int
-        The number of antennas that will be used in beamforming. Each antennas is expected to produce two polarisations.
-    """
-
-    def __init__(self, batches, n_channels, n_blocks, samples_per_block, ants):
-        """Initialise the coefficient generation class."""
-        self.batches = batches
-        self.pols = 2  # Always
-        self.num_chan = n_channels
-        self.n_blocks = n_blocks
-        self.samples_per_block = samples_per_block
-        self.ants = ants
-        self.total_length = self.batches * self.pols * self.num_chan * self.n_blocks * self.samples_per_block
-        self.real_coeff_value = 4
-        self.imag_coeff_value = 1
-
-    @jit
-    def GPU_Coeffs(self):
-        """Generate coefficients for complex multiplication.
-
-        Note: This is for use in complex multiplication using two
-        real-valued arrays. For this reason the coefficients need to be
-        arranged as follows.
-
-        Coefficients Array:
-        [R00  I00
-         -I00 R00
-         R10  I10
-         -I10 R10
-         ...  ...]
-        Where:  R00 = Real coeff 0
-                I00 = Imag coeff 0
-        and the matrix is structured as a N x 2 array.
-
-        Returns
-        -------
-        coeffs: np.ndarray[np.float].
-            Output array of test coefficients.
-        """
-        coeffs = np.empty(((self.ants * 2) * 2 * self.total_length), np.float32).reshape(
-            self.total_length, 2, self.ants * 2
-        )
-        for i in range(coeffs.shape[0]):
-            for j in range(coeffs.shape[1]):
-                for k in range(coeffs.shape[2]):
-                    if j == 0:
-                        if k % 2:
-                            coeffs[i, j, k] = -1 * self.imag_coeff_value
-                        else:
-                            coeffs[i, j, k] = self.real_coeff_value
-                    else:
-                        if k % 2:
-                            coeffs[i, j, k] = self.real_coeff_value
-                        else:
-                            coeffs[i, j, k] = self.imag_coeff_value
-        return coeffs
-
-    @jit
-    def CPU_Coeffs(self):
-        """Generate coefficients for complex multiplication.
-
-        Note: This is for use in complex multiplication using two
-        real-valued arrays. For this reason the coefficients need to be
-        arranged as follows.
-
-        Coefficients Array:
-        [R00  I00
-         -I00 R00
-         R10  I10
-         -I10 R10
-         ...  ...]
-        Where:  R00 = Real coeff 0
-                I00 = Imag coeff 0
-        and the matrix is structured as a N x 2 array.
-
-        Returns
-        -------
-        coeffs: np.ndarray[np.float].
-            Output array of test coefficients.
-        """
-        coeffs = np.empty(self.ants * 2 * self.total_length, np.float32).reshape(self.total_length, self.ants, 2)
-
-        for i in range(coeffs.shape[0]):
-            for j in range(coeffs.shape[1]):
-                for k in range(coeffs.shape[2]):
-                    if k == 0:
-                        coeffs[i, j, k] = self.imag_coeff_value
-                    else:
-                        coeffs[i, j, k] = self.real_coeff_value
-        return coeffs.reshape(
-            self.batches, self.pols, self.num_chan, self.n_blocks, self.samples_per_block, self.ants, 2
-        )
 
 
 @pytest.mark.parametrize("batches", test_parameters.batches)
@@ -172,7 +62,7 @@ def test_beamform_parametrised(batches, num_ants, num_channels, num_samples_per_
     samples_per_block = 16
     n_blocks = num_samples_per_channel // samples_per_block
     coeff_gen = CoeffGenerator(batches, n_channels_per_stream, n_blocks, samples_per_block, num_ants)
-    gpu_coeffs = coeff_gen.GPU_Coeffs()
+    gpu_coeffs = coeff_gen.GPU_Coeffs_cublas()
 
     # 2. Initialise GPU kernels and buffers.
     ctx = accel.create_some_context(device_filter=lambda x: x.is_cuda, interactive=False)
