+++ beamformer/test/beamform_mult_kernel_test.py
@@ -22,118 +22,8 @@ import numpy as np
 import pytest
 import test_parameters
 from beamforming import matrix_multiply
+from coeff_generator import CoeffGenerator
 from katsdpsigproc import accel
-from numba import jit
-
-
-class CoeffGenerator:
-    """Class for generating coefficients for testing purposes.
-
-    Note: The coefficients for this feature will be supplied by CAM and are not
-    generated by CBF.
-
-    Parameters
-    ----------
-    batches: int
-        The number of matrices to be reordered, a single data matrix = one batch.
-    n_channels: int
-        The number of frequency channels to be processed.
-    n_blocks: int
-        The number of blocks that each channels set of samples are divided into.
-    samples_per_block: int
-        The number of time samples to be processed per block.
-    n_ants: int
-        The number of antennas that will be used in beamforming. Each antennas is expected to produce two polarisations.
-    """
-
-    def __init__(self, batches, n_channels, n_blocks, samples_per_block, n_ants):
-        """Initialise the coefficient generation class."""
-        self.batches = batches
-        self.pols = 2  # Always
-        self.num_chan = n_channels
-        self.n_blocks = n_blocks
-        self.samples_per_block = samples_per_block
-        self.n_ants = n_ants
-        self.complexity = 2  # Always
-        self.total_length = self.batches * self.pols * self.num_chan * self.n_blocks * self.samples_per_block
-
-    @jit
-    def GPU_Coeffs(self):
-        """Generate coefficients for complex multiplication on the GPU.
-
-        Note: This is for use in complex multiplication using two
-        real-valued arrays. For this reason the coefficients need to be
-        arranged as follows.
-
-        Coefficients Array:
-        [R00  I00
-         -I00 R00
-         R10  I10
-         -I10 R10
-         ...  ...]
-        Where:  R00 = Real coeff 0
-                I00 = Imag coeff 0
-        and the matrix is structured as a N x 2 array.
-
-        Returns
-        -------
-        coeffs: np.ndarray[np.float32].
-            Output array of test coefficients.
-        """
-        coeffs = np.ones(((self.n_ants * 2) * 2 * self.total_length), np.float32).reshape(
-            self.total_length, 2, self.n_ants * 2
-        )
-        real_value = 4
-        imag_value = 1
-        for i in range(coeffs.shape[0]):
-            for j in range(coeffs.shape[1]):
-                for k in range(coeffs.shape[2]):
-                    if j == 0:
-                        if k % 2:
-                            coeffs[i, j, k] = -1 * imag_value
-                        else:
-                            coeffs[i, j, k] = real_value
-                    else:
-                        if k % 2:
-                            coeffs[i, j, k] = real_value
-                        else:
-                            coeffs[i, j, k] = imag_value
-        return coeffs.reshape(
-            self.batches,
-            self.pols,
-            self.num_chan,
-            self.n_blocks,
-            self.samples_per_block,
-            2,
-            self.n_ants * self.complexity,
-        )
-
-    @jit
-    def CPU_Coeffs(self):
-        """Generate coefficients for complex multiplication on the CPU.
-
-        Note: This is for use in complex multiplication. The real and imaginary
-        values need to be interleaved as <real,imag>
-
-        Returns
-        -------
-        coeffs: np.ndarray[np.float32].
-            Output array of test coefficients.
-        """
-        coeffs = np.ones(self.n_ants * 2 * self.total_length, np.float32).reshape(self.total_length, self.n_ants, 2)
-
-        real_value = 4
-        imag_value = 1
-        for i in range(coeffs.shape[0]):
-            for j in range(coeffs.shape[1]):
-                for k in range(coeffs.shape[2]):
-                    if k == 0:
-                        coeffs[i, j, k] = imag_value
-                    else:
-                        coeffs[i, j, k] = real_value
-        return coeffs.reshape(
-            self.batches, self.pols, self.num_chan, self.n_blocks, self.samples_per_block, self.n_ants, 2
-        )
 
 
 @pytest.mark.parametrize("batches", test_parameters.batches)
@@ -172,7 +62,7 @@ def test_beamform_parametrised(batches, n_ants, num_channels, num_samples_per_ch
     samples_per_block = 16
     n_blocks = num_samples_per_channel // samples_per_block
     coeff_gen = CoeffGenerator(batches, n_channels_per_stream, n_blocks, samples_per_block, n_ants)
-    gpu_coeffs = coeff_gen.GPU_Coeffs()
+    gpu_coeffs = coeff_gen.GPU_Coeffs_kernel()
 
     # 2. Initialise GPU kernels and buffers.
     ctx = accel.create_some_context(device_filter=lambda x: x.is_cuda, interactive=False)
